requirements.txt
google-cloud-bigquery>=3.0.0
PyYAML>=6.0
